+++ CHANGELOG.md
@@ -9,8 +9,11 @@ Changes to the SlideshowFX software are listed by version within this document.
 #### New and noteworthy
 
 * Upgrade FontAwesome to version 5.0.13
-* Upgrade Box hosting connector
-* Upgrade Google Drive hosting connector
+* Upgrade vert.x to version 3.5.2
+* Upgrade Box hosting connector internal library
+* Upgrade Dropbox hosting connector internal library
+* Upgrade Google Drive hosting connector internal library
+* Introducing a presentation outline
 
 #### Bug fixes